Task 2. Execute a command

+++ 0x0A-configuration_management/2-execute_a_command.pp
@@ -2,4 +2,6 @@
 
 exec {'Execute a command':
   command  => 'pkill -f killmenow',
+  path    => ['/usr/bin', '/bin'],
+  onlyif  => 'pgrep -f killmenow',
 }
